[#182] refactor : Change src/routes.ts name and location
	 src/routes.ts -> src/router/apiRoutes.ts
