+++ src/router/index.ts
@@ -1,1 +1,2 @@
 export { Router } from "./router";
+export { kakaoRoutes, userRoutes, partyRoutes, chatRoutes } from "./apiRoutes";
