+++ src/services/internal/auth/http.ts
@@ -1,4 +1,4 @@
-import { userRoutes } from "@/routes";
+import { userRoutes } from "@/router";
 
 import { JJAN_URL } from "../domain";
 import { Response } from "../types";
