+++ src/services/internal/party/query.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { partyRoutes } from "@/routes";
+import { partyRoutes } from "@/router";
 
 import { JJAN_URL } from "../domain";
 import { Response } from "../types";
